chore: fix makefile and gitignore

+++ Makefile
@@ -4,18 +4,10 @@ ROGUEUTIL_HEADER = $(VENDOR_DIR)/rogueutil.h
 MINIAUDIO_URL = https://raw.githubusercontent.com/mackron/miniaudio/master/miniaudio.h
 ROGUEUTIL_URL = https://raw.githubusercontent.com/sakhmatd/rogueutil/master/rogueutil.h
 
-all: fonograf
-
-$(VENDOR_DIR):
-	mkdir -p $(VENDOR_DIR)
-
-$(MINIAUDIO_HEADER): | $(VENDOR_DIR)
-	curl -o $(MINIAUDIO_HEADER) $(MINIAUDIO_URL)
-
-$(ROGUEUTIL_HEADER): | $(VENDOR_DIR)
-	curl -o $(ROGUEUTIL_HEADER) $(ROGUEUTIL_URL)
-
-all: main.cpp fonograf.cpp ui.cpp $(MINIAUDIO_HEADER) $(ROGUEUTIL_HEADER)
+all:
+	mkdir -p $(VENDOR_DIR) || true
+	[ -f $(MINIAUDIO_HEADER) ] || curl -o $(MINIAUDIO_HEADER) $(MINIAUDIO_URL)
+	[ -f $(ROGUEUTIL_HEADER) ] || curl -o $(ROGUEUTIL_HEADER) $(ROGUEUTIL_URL)
 	clang++ -std=c++17 -lpthread -lm -ldl -o fonograf main.cpp fonograf.cpp ui.cpp
 
 .PHONY: clean
